fix the bug

+++ index.cpp
@@ -33,7 +33,7 @@ int main()
               cout << num1 << " * " << num2 << " = " <<(num1 * num2) <<endl;
               break;
               case 4:
-               cout << num1 << " / " << num2 << " = " <<(num1 / num2) <<endl;
+               cout << num1 << " / " << num2 << " = " <<(float)num1 / num2 <<endl;
                break;
         default:
         cout << "ви вели не верний запрос" <<endl;
